+++ lib/grumlin/expressions/scope.rb
@@ -3,11 +3,13 @@
 module Grumlin
   module Expressions
     module Scope
-      extend Tool
-
       SUPPORTED_STEPS = %i[local].freeze
 
-      define_steps(SUPPORTED_STEPS, "Scope")
+      class << self
+        extend Expression
+
+        define_steps(SUPPORTED_STEPS, "Scope")
+      end
     end
   end
 end