+++ lib/grumlin/expressions/u.rb
@@ -4,8 +4,8 @@ module Grumlin
   module Expressions
     module U
       # TODO: add other start steps
-      SUPPORTED_STEPS = %i[V addV count drop fold has hasLabel hasNot id in inE inV is label out outE outV project
-                           repeat select timeLimit unfold valueMap values].freeze
+      SUPPORTED_STEPS = %i[V addV coalesce constant count drop fold has hasLabel hasNot id in inE inV is label out outE
+                           outV project repeat select timeLimit unfold valueMap values].freeze
 
       class << self
         SUPPORTED_STEPS.each do |step|