+++ spec/grumlin/expressions/operator_spec.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+RSpec.describe Grumlin::Expressions::Operator do
+  include_examples "has sorted constant", "SUPPORTED_STEPS"
+end